+++ src/main.js
@@ -4,11 +4,18 @@ import Vue from 'vue'
 import App from './App'
 import router from './router'
 import VueVideoPlayer from 'vue-video-player'
+// videojs
+import videojs from 'video.js'
+
 import 'video.js/dist/video-js.css'
 import 'vue-video-player/src/custom-theme.css'
 import 'videojs-flash'
-import 'videojs-contrib-hls/dist/videojs-contrib-hls'
-// import 'videojs-contrib-hls.js/src/videojs.hlsjs'
+
+window.videojs = videojs
+
+// 由于 import 必须置顶，此句必须 使用 window.videojs, 所以这里用 require
+require('videojs-contrib-hls/dist/videojs-contrib-hls')
+// require('videojs-contrib-hls.js/src/videojs.hlsjs')
 
 Vue.config.productionTip = false
 Vue.use(VueVideoPlayer)