+++ CHANGELOG.md
@@ -1,17 +1,28 @@
 # v1.0.1
+
 - RTMP & HLS support
 
 # v1.0.2
+
 - Normal playback support in `/playback`
 
 # v1.0.3
+
 - use bootstrap4
 - update vue to 2.5.13
 
 # v1.0.4
+
 - upgrade bootstrap 4
 - ga
 
 # v1.0.5
+
 - npm update
 - use local swf
+
+# v2.0.0
+
+- update vue-video-player to v5.0.2
+- update vue-router to v3.0.1
+- update videojs-flash to v2.1.1